♻️ celery 코드 일부 삭제밎 리팩토링
