+++ app/celery/schema.py
@@ -2,8 +2,8 @@ import strawberry
 from strawberry import Schema
 from strawberry.tools import merge_types
 
-import app.celery.comment.schema
-import app.celery.count.schema
+import app.routes.count.count_controller
+import app.routes.comment.comment_comtroller
 import app.celery.LLM.schema
 import app.routes.board.board_controller
 
@@ -11,22 +11,23 @@ Query = merge_types(
     "Query",
     (
         app.celery.LLM.schema.Query,
-        app.celery.comment.schema.Query,
+        app.routes.comment.comment_comtroller.Query,
         app.routes.board.board_controller.Query,
-        app.celery.count.schema.Query,
+        app.routes.count.count_controller.Query,
     ),
 )
 status_query = merge_types(
     "Query",
     (app.celery.LLM.schema.TaskStatusQuery,
-     app.celery.comment.schema.TaskStatusQuery),
+     )
+     # app.celery.comment.schema.TaskStatusQuery),
 )
 Mutation = merge_types(
     "Mutation",
     (
         app.celery.LLM.schema.Mutation,
-        app.celery.comment.schema.Mutation,
-        app.celery.count.schema.Mutation,
+        app.routes.comment.comment_comtroller.Mutation,
+        app.routes.count.count_controller.Mutation,
     ),
 )
 schema = Schema(query=Query, mutation=Mutation)
